gpu: pvr: pdumpfs: add stream_frames debugfs entry
A mechanism to read out this data reliably is put in place. And data
is only lost when the stream is no longer being read or when a hard
limit of 1024 frames has been reached when the stream is still open.
In case of missing data, the buffer is padded with "..." to make
up for the missing bytes and to clear mark them.

Signed-off-by: Luc Verhaegen <libv@codethink.co.uk>
Signed-off-by: Imre Deak <imre.deak@nokia.com>